This is the big one.
Bring in support for a rpm based installation of chef server/client via
bootstrap. This is done by adding a dist_only? attribute that, if passed to the
bootstrap, would attempt to install packages via yum as opposed to the current
gem+deb install.

This commit also brings in init scripts for chef-client and chef-indexer as
opposed to carrying them in rpms.